fett

+++ GE/worldStructFunctions.go
@@ -36,7 +36,7 @@ func GetWorldStructureFromBytes(X,Y,W,H float64, data []byte, tile_path, struct_
 	p.BytesToObjects(bs[6])
 	p.BytesToLights(bs[7])
 	p.SetMiddle(int(cmp.BytesToInt64(bs[0])), int(cmp.BytesToInt64(bs[1])), true)
-	p.SetLightStats(int64(cmp.BytesToInt16(bs[2])), int64(cmp.BytesToInt16(bs[3])), cmp.BytesToFloat64(bs[4]))
+	p.SetLightStats(int16(cmp.BytesToInt16(bs[2])), int16(cmp.BytesToInt16(bs[3])), cmp.BytesToFloat64(bs[4]))
 	return p, nil
 }
